Move and rename challenge package
