+++ src/main/java/de/jilence/jutils/challenges/RandomEffectOnDamage.java
@@ -1,22 +1,26 @@
-package de.jilence.jutils.challenge.challenge;
+package de.jilence.jutils.challenges;
 
 import de.jilence.jutils.Main;
 import de.jilence.jutils.challenge.Challenge;
 import de.jilence.jutils.challenge.ChallengeManager;
 import de.jilence.jutils.utils.ItemBuilder;
 import de.jilence.jutils.utils.LoreBuilder;
-import net.kyori.adventure.text.Component;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import org.bukkit.event.block.BlockPlaceEvent;
+import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.event.inventory.ClickType;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.potion.PotionEffect;
+import org.bukkit.potion.PotionEffectType;
 
-public class NoBlockPlace extends Challenge implements Listener {
+import java.util.Random;
+
+public class RandomEffectOnDamage extends Challenge implements Listener {
+
     @Override
     public void onEnable() {
 
@@ -24,7 +28,7 @@ public class NoBlockPlace extends Challenge implements Listener {
 
     @Override
     public void onStart() {
-        Bukkit.getPluginManager().registerEvents(new NoBlockPlace(), Main.getPlugin(Main.class));
+        Bukkit.getPluginManager().registerEvents(new RandomEffectOnDamage(), Main.getPlugin(Main.class));
     }
 
     @Override
@@ -37,15 +41,26 @@ public class NoBlockPlace extends Challenge implements Listener {
 
     }
 
-    @Override
-    public void onTick() {
-        Bukkit.broadcast(Component.text("TICK"), "bukkit.broadcast");
+    @EventHandler(ignoreCancelled = true)
+    public void onEntityDamage(EntityDamageEvent event) {
+
+        if(event.getEntity() instanceof Player) {
+
+            Random random = new Random();
+            int i = random.nextInt(((PotionEffectType.values().length - 1)) + 1);
+            PotionEffectType effect = PotionEffectType.values()[i];
+
+            Bukkit.getOnlinePlayers().forEach(player -> player.addPotionEffect(new PotionEffect(effect, 99999, 1)));
+
+        }
+
     }
 
-    @EventHandler(ignoreCancelled = true)
-    public void onBlockPlace(BlockPlaceEvent event) {
-        event.getPlayer().sendMessage("SDasdasdasd");
+    @Override
+    public void onTick() {
+
     }
+
 
     @Override
     public Inventory getSettingsInventory() {
@@ -54,18 +69,18 @@ public class NoBlockPlace extends Challenge implements Listener {
 
     @Override
     public ItemStack getDisplayItem() {
-        ItemStack itemStack = new ItemBuilder(Material.GRASS_BLOCK, 1).displayname("§7No-Block-Place §9Challenge").build();
+        ItemStack itemStack = new ItemBuilder(Material.POTION, 1).displayname("§7Effect on Damage §9Challenge").build();
         new LoreBuilder(itemStack)
                 .addSpace()
                 .addDescription(LoreBuilder.DESCRIPTIONS.HEADER)
                 .addSpace()
-                .addLoreLine("§7Bei Dieser §9Herausforderung §7darf kein Spieler,")
-                .addLoreLine("§7einen Block platzieren")
+                .addLoreLine("§7Wenn ein §9Spieler §7Schaden bekommt,")
+                .addLoreLine("§7bekommen alle §9Spieler §7einen random Effect.")
                 .addSpace()
                 .addDescription(LoreBuilder.DESCRIPTIONS.RIGHT_LEFT_CLICK)
                 .addSpace()
                 .addDescription(LoreBuilder.DESCRIPTIONS.SETTINGS)
-                .addBoolSettings(ChallengeManager.isChallengeEnabled(ChallengeManager.Challenges.NOBLOCKPLACE), "§7Herausforderung")
+                .addBoolSettings(ChallengeManager.isChallengeEnabled(ChallengeManager.Challenges.RANDOMEFFECTONDAMAGE), "§7Herausforderung")
                 .addSpace();
 
         return itemStack;
@@ -80,5 +95,4 @@ public class NoBlockPlace extends Challenge implements Listener {
     public void onInventoryClick(Player player, ClickType clickType, Material material, Inventory inventory) {
 
     }
-
 }
